Move JLR*. to shifters move sboxes back to parent
