+++ src/stm32_vcu.cpp
@@ -42,6 +42,7 @@
 #include "bms/oibms.h"
 #include "bms/simpbms.h"
 #include "bms/stwmbms.h"
+#include "bmw_sbox.h"
 #include "canmap.h"
 #include "cansdo.h"
 #include "chargers/C5Charger.h"
@@ -86,6 +87,8 @@
 #include "printf.h"
 #include "shifters/E65_Lever.h"
 #include "shifters/F30_Lever.h"
+#include "shifters/JLR_G1.h"
+#include "shifters/JLR_G2.h"
 #include "shifters/PKP2300_Lever.h"
 #include "shifters/shifter.h"
 #include "stm32_can.h"
@@ -95,16 +98,13 @@
 #include "terminalcommands.h"
 #include "throttle.h"
 #include "utils.h"
+#include "vag_sbox.h"
 #include "vehicles/BMW_E31.h"
 #include "vehicles/BMW_E39.h"
 #include "vehicles/BMW_E65.h"
 #include "vehicles/Can_VAG.h"
-#include "vehicles/JLR_G1.h"
-#include "vehicles/JLR_G2.h"
 #include "vehicles/V_Classic.h"
-#include "vehicles/bmw_sbox.h"
 #include "vehicles/subaruvehicle.h"
-#include "vehicles/vag_sbox.h"
 #include "vehicles/vehicle.h"
 #include <libopencm3/stm32/can.h>
 #include <libopencm3/stm32/exti.h>
